Add tests to OverrideFileldValueValidator
Minor: reorder imports and move to proper module
